modify strjoin, size_t ft

+++ libft/ft_strlcat.c
@@ -6,13 +6,13 @@
 /*   By: ylee <ylee@student.42seoul.kr>             +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/10/07 11:07:51 by ylee              #+#    #+#             */
-/*   Updated: 2020/10/12 13:45:55 by ylee             ###   ########.fr       */
+/*   Updated: 2020/10/13 13:03:38 by ylee             ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include <stddef.h>
 
-int		ft_strlcat(char *dst, const char *src, size_t size)
+size_t	ft_strlcat(char *dst, const char *src, size_t size)
 {
 	size_t	i;
 	size_t	dst_len;
